[ModelicaSystem] split file
